changed datatype of balance to double

+++ library/src/main/java/com/thoughtworks/step_bank/Account.java
@@ -2,10 +2,10 @@ package com.thoughtworks.step_bank;
 
 public class Account {
     private final String accountNumber;
-    private int balance;
+    private double balance;
 
 
-    public Account(String accountNumber, int balance) throws MinimumBalanceException, InvalidAccountNumberException {
+    public Account(String accountNumber, double balance) throws MinimumBalanceException, InvalidAccountNumberException {
         if (incorrectAccountNumberPattern(accountNumber)) {
             throw new InvalidAccountNumberException();
         }
@@ -24,16 +24,16 @@ public class Account {
         return accountNumber;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public int credit(int amount) {
+    public double credit(int amount) {
         balance+=amount;
         return balance;
     }
 
-    public int debit(int amount) throws MinimumBalanceException{
+    public double debit(int amount) throws MinimumBalanceException{
         if (balance-amount<100){
             throw new MinimumBalanceException();
         }
